Migrate collector to component implementation
