+++ cmd/cluster-agent-cloudfoundry/subcommands/run/command.go
@@ -146,8 +146,7 @@ func run(log log.Component, taggerComp tagger.Component, demultiplexer demultipl
 	common.LoadComponents(secretResolver, wmeta, pkgconfig.Datadog.GetString("confd_path"))
 
 	// Set up check collector
-	common.AC.AddScheduler("check", pkgcollector.InitCheckScheduler(optional.NewOption[pkgcollector.Collector](collector), demultiplexer), true)
-	collector.Start()
+	common.AC.AddScheduler("check", pkgcollector.InitCheckScheduler(optional.NewOption(collector), demultiplexer), true)
 	diagnose.Init(optional.NewOption(collector))
 
 	// start the autoconfig, this will immediately run any configured check
