+++ comp/collector/collector/collectorimpl/collector.go
@@ -7,18 +7,34 @@
 package collectorimpl
 
 import (
+	"context"
+	"fmt"
+	"sync"
+	"time"
+
+	"go.uber.org/atomic"
 	"go.uber.org/fx"
 
 	"github.com/DataDog/datadog-agent/cmd/agent/common"
-	"github.com/DataDog/datadog-agent/comp/aggregator/demultiplexer"
 	"github.com/DataDog/datadog-agent/comp/collector/collector"
+	"github.com/DataDog/datadog-agent/comp/collector/collector/collectorimpl/internal/middleware"
 	"github.com/DataDog/datadog-agent/comp/core/config"
 	"github.com/DataDog/datadog-agent/comp/core/log"
 	"github.com/DataDog/datadog-agent/comp/core/status"
-	pkgcollector "github.com/DataDog/datadog-agent/pkg/collector"
+	"github.com/DataDog/datadog-agent/pkg/aggregator/sender"
+	"github.com/DataDog/datadog-agent/pkg/collector/check"
+	checkid "github.com/DataDog/datadog-agent/pkg/collector/check/id"
+	"github.com/DataDog/datadog-agent/pkg/collector/runner"
+	"github.com/DataDog/datadog-agent/pkg/collector/runner/expvars"
+	"github.com/DataDog/datadog-agent/pkg/collector/scheduler"
 	collectorStatus "github.com/DataDog/datadog-agent/pkg/status/collector"
 	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
 	"github.com/DataDog/datadog-agent/pkg/util/optional"
+)
+
+const (
+	stopped uint32 = iota
+	started
 )
 
 type dependencies struct {
@@ -28,11 +44,26 @@ type dependencies struct {
 	Config config.Component
 	Log    log.Component
 
-	Demultiplexer demultiplexer.Component
+	SenderManager sender.SenderManager
 }
 
 type collectorImpl struct {
-	pkgcollector.Collector
+	log log.Component
+
+	senderManager  sender.SenderManager
+	checkInstances int64
+
+	// state is 'started' or 'stopped'
+	state *atomic.Uint32
+
+	scheduler      *scheduler.Scheduler
+	runner         *runner.Runner
+	checks         map[checkid.ID]*middleware.CheckWrapper
+	eventReceivers []collector.EventReceiver
+
+	cancelCheckTimeout time.Duration
+
+	m sync.RWMutex
 }
 
 type provides struct {
@@ -46,7 +77,7 @@ type provides struct {
 // Module defines the fx options for this component.
 func Module() fxutil.Module {
 	return fxutil.Component(
-		fx.Provide(newCollector),
+		fx.Provide(newProvides),
 	)
 }
 
@@ -57,11 +88,8 @@ func ModuleNoneCollector() fxutil.Module {
 	)
 }
 
-func newCollector(deps dependencies) provides {
-	c := &collectorImpl{
-		Collector: pkgcollector.NewCollector(deps.Demultiplexer, deps.Config.GetDuration("check_cancel_timeout"), common.GetPythonPaths()...),
-	}
-
+func newProvides(deps dependencies) provides {
+	c := newCollector(deps)
 	return provides{
 		Comp:         c,
 		OptionalComp: optional.NewOption[collector.Component](c),
@@ -74,4 +102,267 @@ func newNoneCollector() provides {
 		OptionalComp: optional.NewNoneOption[collector.Component](),
 		Provider:     status.NewInformationProvider(collectorStatus.Provider{}),
 	}
+}
+
+func newCollector(deps dependencies) *collectorImpl {
+	c := &collectorImpl{
+		log:                deps.Log,
+		senderManager:      deps.SenderManager,
+		checks:             make(map[checkid.ID]*middleware.CheckWrapper),
+		state:              atomic.NewUint32(stopped),
+		checkInstances:     int64(0),
+		cancelCheckTimeout: deps.Config.GetDuration("check_cancel_timeout"),
+	}
+	pyVer, pyHome, pyPath := pySetup(common.GetPythonPaths()...)
+
+	// print the Python info if the interpreter was embedded
+	if pyVer != "" {
+		c.log.Infof("Embedding Python %s", pyVer)
+		c.log.Debugf("Python Home: %s", pyHome)
+		c.log.Debugf("Python path: %s", pyPath)
+	}
+
+	// Prepare python environment if necessary
+	if err := pyPrepareEnv(); err != nil {
+		c.log.Errorf("Unable to perform additional configuration of the python environment: %v", err)
+	}
+
+	deps.Lc.Append(fx.Hook{
+		OnStart: c.start,
+		OnStop:  c.stop,
+	})
+
+	c.log.Debug("Collector up and running!")
+	return c
+}
+
+// AddEventReceiver adds a callback to the collector to be called each time a check is added or removed.
+func (c *collectorImpl) AddEventReceiver(cb collector.EventReceiver) {
+	c.m.Lock()
+	defer c.m.Unlock()
+
+	c.eventReceivers = append(c.eventReceivers, cb)
+}
+
+func (c *collectorImpl) notify(cid checkid.ID, e collector.EventType) {
+	for _, cb := range c.eventReceivers {
+		cb(cid, e)
+	}
+}
+
+// start begins the collector's operation.  The scheduler will not run any checks until this has been called.
+func (c *collectorImpl) start(_ context.Context) error {
+	c.m.Lock()
+	defer c.m.Unlock()
+
+	run := runner.NewRunner(c.senderManager)
+	sched := scheduler.NewScheduler(run.GetChan())
+
+	// let the runner some visibility into the scheduler
+	run.SetScheduler(sched)
+	sched.Run()
+
+	c.scheduler = sched
+	c.runner = run
+	c.state.Store(started)
+
+	return nil
+}
+
+// stop halts any component involved in running a Check
+func (c *collectorImpl) stop(_ context.Context) error {
+	c.m.Lock()
+	defer c.m.Unlock()
+
+	if c.scheduler != nil {
+		c.scheduler.Stop() //nolint:errcheck
+		c.scheduler = nil
+	}
+	if c.runner != nil {
+		c.runner.Stop()
+		c.runner = nil
+	}
+	c.state.Store(stopped)
+	return nil
+}
+
+// RunCheck sends a Check in the execution queue
+func (c *collectorImpl) RunCheck(inner check.Check) (checkid.ID, error) {
+	c.m.Lock()
+	defer c.m.Unlock()
+
+	ch := middleware.NewCheckWrapper(inner, c.senderManager)
+
+	var emptyID checkid.ID
+
+	if c.state.Load() != started {
+		return emptyID, fmt.Errorf("the collector is not running")
+	}
+
+	if _, found := c.checks[ch.ID()]; found {
+		return emptyID, fmt.Errorf("a check with ID %s is already running", ch.ID())
+	}
+
+	err := c.scheduler.Enter(ch)
+	if err != nil {
+		return emptyID, fmt.Errorf("unable to schedule the check: %s", err)
+	}
+
+	// Track the total number of checks running in order to have an appropriate number of workers
+	c.checkInstances++
+	if ch.Interval() == 0 {
+		// Adding a temporary runner for long running check in case the
+		// number of runners is lower than the number of long running
+		// checks.
+		c.log.Infof("Adding an extra runner for the '%s' long running check", ch)
+		c.runner.AddWorker()
+	} else {
+		c.runner.UpdateNumWorkers(c.checkInstances)
+	}
+
+	c.checks[ch.ID()] = ch
+	c.notify(ch.ID(), collector.CheckRun)
+	return ch.ID(), nil
+}
+
+// StopCheck halts a check and remove the instance
+func (c *collectorImpl) StopCheck(id checkid.ID) error {
+	if !c.started() {
+		return fmt.Errorf("the collector is not running")
+	}
+
+	ch, found := c.get(id)
+	if !found {
+		return fmt.Errorf("cannot find a check with ID %s", id)
+	}
+
+	// unschedule the instance
+	err := c.scheduler.Cancel(id)
+	if err != nil {
+		return fmt.Errorf("an error occurred while canceling the check schedule: %s", err)
+	}
+
+	err = c.runner.StopCheck(id)
+	if err != nil {
+		// still attempt to cancel the check before returning the error
+		_ = c.cancelCheck(ch, c.cancelCheckTimeout)
+		return fmt.Errorf("an error occurred while stopping the check: %s", err)
+	}
+
+	err = c.cancelCheck(ch, c.cancelCheckTimeout)
+	if err != nil {
+		return fmt.Errorf("an error occurred while calling check.Cancel(): %s", err)
+	}
+
+	// remove the check from the stats map
+	expvars.RemoveCheckStats(id)
+
+	// vaporize the check
+	c.delete(id)
+
+	return nil
+}
+
+// cancelCheck calls Cancel on the passed check, with a timeout
+func (c *collectorImpl) cancelCheck(ch check.Check, timeout time.Duration) error {
+	done := make(chan struct{})
+
+	go func() {
+		ch.Cancel()
+		close(done)
+	}()
+
+	select {
+	case <-done:
+		return nil
+	case <-time.After(timeout):
+		return fmt.Errorf("timeout while calling check.Cancel() on check ID %s, timeout: %s", ch.ID(), timeout)
+	}
+}
+
+func (c *collectorImpl) get(id checkid.ID) (check.Check, bool) {
+	c.m.RLock()
+	defer c.m.RUnlock()
+
+	ch, found := c.checks[id]
+	return ch, found
+}
+
+// remove the check from the list
+func (c *collectorImpl) delete(id checkid.ID) {
+	c.m.Lock()
+	defer c.m.Unlock()
+
+	delete(c.checks, id)
+	c.notify(id, collector.CheckStop)
+}
+
+// lightweight shortcut to see if the collector has started
+func (c *collectorImpl) started() bool {
+	return c.state.Load() == started
+}
+
+// MapOverChecks call the callback with the list of checks locked.
+func (c *collectorImpl) MapOverChecks(cb func([]check.Info)) {
+	c.m.RLock()
+	defer c.m.RUnlock()
+
+	cInfo := []check.Info{}
+	for _, c := range c.checks {
+		cInfo = append(cInfo, c)
+	}
+	cb(cInfo)
+}
+
+// GetChecks copies checks
+func (c *collectorImpl) GetChecks() []check.Check {
+	c.m.RLock()
+	defer c.m.RUnlock()
+
+	chks := make([]check.Check, 0, len(c.checks))
+	for _, chck := range c.checks {
+		chks = append(chks, chck)
+	}
+
+	return chks
+}
+
+// GetAllInstanceIDs returns the ID's of all instances of a check
+func (c *collectorImpl) GetAllInstanceIDs(checkName string) []checkid.ID {
+	c.m.RLock()
+	defer c.m.RUnlock()
+
+	instances := []checkid.ID{}
+	for id, check := range c.checks {
+		if check.String() == checkName {
+			instances = append(instances, id)
+		}
+	}
+
+	return instances
+}
+
+// ReloadAllCheckInstances completely restarts a check with a new configuration
+func (c *collectorImpl) ReloadAllCheckInstances(name string, newInstances []check.Check) ([]checkid.ID, error) {
+	if !c.started() {
+		return nil, fmt.Errorf("The collector is not running")
+	}
+
+	// Stop all the old instances
+	killed := c.GetAllInstanceIDs(name)
+	for _, id := range killed {
+		e := c.StopCheck(id)
+		if e != nil {
+			return nil, fmt.Errorf("Error stopping check %s: %s", id, e)
+		}
+	}
+
+	// Start the new instances
+	for _, check := range newInstances {
+		id, e := c.RunCheck(check)
+		if e != nil {
+			return nil, fmt.Errorf("Error adding check %s: %s", id, e)
+		}
+	}
+	return killed, nil
 }
