+++ pkg/collector/scheduler.go
@@ -3,6 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
+// Package collector provides the implementation of the collector
 package collector
 
 import (
@@ -14,6 +15,7 @@ import (
 	"golang.org/x/text/cases"
 	"golang.org/x/text/language"
 
+	"github.com/DataDog/datadog-agent/comp/collector/collector"
 	"github.com/DataDog/datadog-agent/pkg/aggregator/sender"
 	"github.com/DataDog/datadog-agent/pkg/autodiscovery/integration"
 	"github.com/DataDog/datadog-agent/pkg/collector/check"
@@ -54,13 +56,13 @@ func init() {
 type CheckScheduler struct {
 	configToChecks map[string][]checkid.ID // cache the ID of checks we load for each config
 	loaders        []check.Loader
-	collector      optional.Option[Collector]
+	collector      optional.Option[collector.Component]
 	senderManager  sender.SenderManager
 	m              sync.RWMutex
 }
 
 // InitCheckScheduler creates and returns a check scheduler
-func InitCheckScheduler(collector optional.Option[Collector], senderManager sender.SenderManager) *CheckScheduler {
+func InitCheckScheduler(collector optional.Option[collector.Component], senderManager sender.SenderManager) *CheckScheduler {
 	checkScheduler = &CheckScheduler{
 		collector:      collector,
 		senderManager:  senderManager,
@@ -137,12 +139,8 @@ func (s *CheckScheduler) Unschedule(configs []integration.Config) {
 	}
 }
 
-// Stop handles clean stop of registered schedulers
-func (s *CheckScheduler) Stop() {
-	if coll, ok := s.collector.Get(); ok {
-		coll.Stop()
-	}
-}
+// Stop is a stub to satisfy the scheduler interface
+func (s *CheckScheduler) Stop() {}
 
 // AddLoader adds a new Loader that AutoConfig can use to load a check.
 func (s *CheckScheduler) AddLoader(loader check.Loader) {
